Various updates
- added drop behaviour
- fixed major ANCOVA issues
- improved definition syntax
