+++ jamovi/ui/ancova.src.js
@@ -2,8 +2,7 @@
 'use strict';
 
 const options = require('./ancova.options');
-
-const controls = new ControlManager();
+const actions = require('./ancova.actions');
 
 const ancovaLayout = ui.extend({
 
@@ -24,8 +23,9 @@ const ancovaLayout = ui.extend({
                     label: "Dependent Variable",
                     showColumnHeaders: false,
                     maxItemCount: 1,
+                    itemDropBehaviour: "overwrite",
                     columns: [
-                        { type: "listitem.variablelabel", name: "column1", label: "", format: FormatDef.variable, stretchFactor: 1 }
+                        { type: "listitem.variablelabel", name: "column1", label: "", stretchFactor: 1 }
                     ]
                 },
                 {
@@ -34,7 +34,7 @@ const ancovaLayout = ui.extend({
                     label: "Fixed Factors",
                     showColumnHeaders: false,
                     columns: [
-                        { type: "listitem.variablelabel", name: "column1", label: "", format: FormatDef.variable, stretchFactor: 1 }
+                        { type: "listitem.variablelabel", name: "column1", label: "", stretchFactor: 1 }
                     ]
                 },
                 {
@@ -43,25 +43,16 @@ const ancovaLayout = ui.extend({
                     label: "Covariates",
                     showColumnHeaders: false,
                     columns: [
-                        { type: "listitem.variablelabel", name: "column1", label: "", format: FormatDef.variable, stretchFactor: 1 }
-                    ]
-                },
-                {
-                    name: "wlsWeights",
-                    type:"variabletargetlistbox",
-                    label: "WLS Weights",
-                    showColumnHeaders: false,
-                    maxItemCount: 1,
-                    columns: [
-                        { type: "listitem.variablelabel", name: "column1", label: "", format: FormatDef.variable, stretchFactor: 1 }
+                        { type: "listitem.variablelabel", name: "column1", label: "", stretchFactor: 1 }
                     ]
                 }
             ]
         },
         {
-            type: "layoutbox",
+            type: "label",
+            label: "Effect Size",
+            style: "list-inline",
             margin: "large",
-            style: "inline",
             controls: [
                 { name: "etaSq",   type:"checkbox", label: "η²" },
                 { name: "etaSqP",  type:"checkbox", label: "partial η²" },
@@ -87,8 +78,9 @@ const ancovaLayout = ui.extend({
                             label: "Model Terms",
                             showColumnHeaders: false,
                             valueFilter: "unique",
+                            itemDropBehaviour: "empty_space",
                             columns: [
-                                { type: "listitem.variablelabel", name: "column1", label: "", format: FormatDef.variable, stretchFactor: 1 }
+                                { type: "listitem.termlabel", name: "column1", label: "", stretchFactor: 1 }
                             ]
                         }
                     ]
@@ -186,6 +178,7 @@ const ancovaLayout = ui.extend({
                             label: "Horizontal axis",
                             showColumnHeaders: false,
                             maxItemCount: 1,
+                            itemDropBehaviour: "overwrite",
                             columns: [
                                 { type: "listitem.variablelabel", name: "column1", label: "", format: FormatDef.variable, stretchFactor: 1 }
                             ]
@@ -196,6 +189,7 @@ const ancovaLayout = ui.extend({
                             label: "Separate lines",
                             showColumnHeaders: false,
                             maxItemCount: 1,
+                            itemDropBehaviour: "overwrite",
                             columns: [
                                 { type: "listitem.variablelabel", name: "column1", label: "", format: FormatDef.variable, stretchFactor: 1 }
                             ]
@@ -206,6 +200,7 @@ const ancovaLayout = ui.extend({
                             label: "Separate plots",
                             showColumnHeaders: false,
                             maxItemCount: 1,
+                            itemDropBehaviour: "overwrite",
                             columns: [
                                 { type: "listitem.variablelabel", name: "column1", label: "", format: FormatDef.variable, stretchFactor: 1 }
                             ]
@@ -232,23 +227,6 @@ const ancovaLayout = ui.extend({
             stretchFactor: 1,
             controls : [
                 {
-                    type: "supplier",
-                    name: "marginalMeansSupplier",
-                    stretchFactor: 1,
-                    persistentItems: false,
-                    controls: [
-                        {
-                            type:"targetlistbox",
-                            name: "margMeans",
-                            label: "Marginal means",
-                            showColumnHeaders: false,
-                            columns: [
-                                { type: "listitem.variablelabel", name: "column1", label: "", format: FormatDef.variable, stretchFactor: 1 }
-                            ]
-                        }
-                    ]
-                },
-                {
                     name: "compMain", label: "Compare main effects", type:"checkbox",
                     controls: [
                         { name: "compMainCorr", type:"combobox", label: "Correction", options: [
@@ -269,292 +247,6 @@ const ancovaLayout = ui.extend({
             ]
         }
     ],
-
-    actions: [
-        {
-            onChange: "compMain", execute: function(ui) {
-                ui.compMainCorr.setEnabled(ui.compMain.value());
-            }
-        },
-        {
-            onChange: "plotError", execute: function(ui) {
-                ui.ciWidth.setEnabled(ui.plotError.value() === "ci");
-            }
-        },
-        {
-            onChange: ["fixedFactors", "covariates"], execute: function(ui) {
-                this.calcModelTerms(ui);
-            }
-        },
-        {
-            onChange: "modelTerms", execute: function(ui) {
-                this.filterModelTerms(ui);
-            }
-        },
-        {
-            onEvent: "view.data-initialising", execute: function(ui) {
-                this._lastVariableList = null;
-                this._lastCurrentList = null;
-                this._lastCombinedList = null;
-                this._lastCovariatesList = null;
-                this._initialising = true;
-            }
-        },
-        {
-            onEvent: "view.data-initialised", execute: function(ui) {
-                if (this._lastVariableList === null || this._lastCombinedList === null || this._lastCovariatesList === null)
-                    this.calcModelTerms(ui);
-
-                if (this._lastCurrentList === null)
-                    this.filterModelTerms(ui);
-
-                this._initialising = false;
-            }
-        }
-    ],
-
-    filterModelTerms: function(ui) {
-        var currentList = this.clone(ui.modelTerms.value());
-        if (currentList === null)
-            currentList = [];
-
-        var covariatesList = ui.covariates.value();
-        if (covariatesList === null)
-            covariatesList = [];
-
-        var covariateFreeList = [];
-        for (let i = 0; i < currentList.length; i++) {
-            let newVar = currentList[i];
-            if (this.containsCovariate(newVar, covariatesList) === false)
-                covariateFreeList.push(this.clone(newVar));
-        }
-
-        var list = this.convertArrayToSupplierList(covariateFreeList, FormatDef.variable);
-        ui.marginalMeansSupplier.setValue(list);
-
-        var diff = null;
-        if ( ! this._initialising)
-            diff = this.findDifferences(this._lastCurrentList, currentList);
-        this._lastCurrentList = currentList;
-
-        if (this._initialising)
-            return;
-
-        var changed = false;
-        if (diff.removed.length > 0 && currentList !== null) {
-            var itemsRemoved = false;
-            for (var i = 0; i < diff.removed.length; i++) {
-                var item = diff.removed[i];
-                for (var j = 0; j < currentList.length; j++) {
-                    if (FormatDef.variable.contains(currentList[j], item)) {
-                        currentList.splice(j, 1);
-                        j -= 1;
-                        itemsRemoved = true;
-                    }
-                }
-            }
-
-            if (itemsRemoved)
-                changed = true;
-        }
-
-        if (this.sortByLength(currentList))
-            changed = true;
-
-        if (changed)
-            ui.modelTerms.setValue(currentList);
-
-        if (diff.removed.length > 0) {
-            itemsRemoved = false;
-            var margMeans = this.clone(ui.margMeans.value());
-            if (margMeans === null)
-                margMeans = [];
-
-            for (var i = 0; i < diff.removed.length; i++) {
-                var item = diff.removed[i];
-                for (var j = 0; j < margMeans.length; j++) {
-                    if (FormatDef.variable.contains(margMeans[j], item)) {
-                        margMeans.splice(j, 1);
-                        j -= 1;
-                        itemsRemoved = true;
-                    }
-                }
-            }
-
-            if (itemsRemoved) {
-                ui.margMeans.setValue(margMeans);
-            }
-        }
-    },
-
-    calcModelTerms : function(ui) {
-        var variableList = this.clone(ui.fixedFactors.value());
-        if (variableList === null)
-            variableList = [];
-
-        var covariatesList = this.clone(ui.covariates.value());
-        if (covariatesList === null)
-            covariatesList = [];
-
-        var combinedList = variableList.concat(covariatesList);
-
-        ui.modelSupplier.setValue(this.convertArrayToSupplierList(combinedList, FormatDef.variable));
-        ui.plotsSupplier.setValue(this.convertArrayToSupplierList(variableList, FormatDef.variable));
-        ui.postHocSupplier.setValue(this.convertArrayToSupplierList(variableList, FormatDef.variable));
-
-        var diff = { removed: [], added: [] };
-        if (this._lastVariableList !== null)
-            diff = this.findDifferences(this._lastVariableList, variableList);
-        this._lastVariableList = variableList;
-
-        var diff2 = { removed: [], added: [] };
-        if (this._lastCovariatesList !== null)
-            diff2 = this.findDifferences(this._lastCovariatesList, covariatesList);
-        this._lastCovariatesList = covariatesList;
-
-        var combinedDiff = { removed: [], added: [] };
-        if (this._lastCombinedList !== null)
-            combinedDiff = this.findDifferences(this._lastCombinedList, combinedList);
-        this._lastCombinedList = combinedList;
-
-        if (this._initialising)
-            return;
-
-        var currentList = this.clone(ui.modelTerms.value());
-        if (currentList === null)
-            currentList = [];
-
-        for (var i = 0; i < combinedDiff.removed.length; i++) {
-            for (var j = 0; j < currentList.length; j++) {
-                if (FormatDef.variable.contains(currentList[j], diff.removed[i])) {
-                    currentList.splice(j, 1);
-                    j -= 1;
-                }
-            }
-        }
-
-        if (currentList === null)
-            currentList = [];
-
-        for (var i = 0; i < diff.added.length; i++) {
-            var listLength = currentList.length;
-            for (var j = 0; j < listLength; j++) {
-                var newVar = currentList[j];
-                if (this.containsCovariate(newVar, covariatesList) === false) {
-                    if (Array.isArray(newVar))
-                        newVar = this.clone(newVar);
-                    else
-                        newVar = [newVar];
-                    newVar.push(diff.added[i])
-                    currentList.push(newVar);
-                }
-            }
-            currentList.push(diff.added[i]);
-        }
-
-        for (var i = 0; i < diff2.added.length; i++)
-            currentList.push(diff2.added[i]);
-
-        ui.modelTerms.setValue(currentList);
-
-        this.updateContrasts(ui, variableList);
-    },
-
-    updateContrasts : function(ui, variableList) {
-        var currentList = this.clone(ui.contrasts.value());
-        if (currentList === null)
-            currentList = [];
-
-        var list3 = [];
-        for (var i = 0; i < variableList.length; i++)
-            list3.push({ var: variableList[i], type: "none" });
-
-        ui.contrasts.setValue(list3);
-    },
-
-    sortByLength : function(list) {
-        var changed = false;
-        for (var i = 0; i < list.length; i++) {
-            var l1 = 1;
-            if (Array.isArray(list[i]))
-                l1 = list[i].length;
-
-            var l2 = 1;
-            if (Array.isArray(list[i+1]))
-                l2 = list[i+1].length;
-
-            if (list.length > i + 1 && (l1 > l2)) {
-                changed = true;
-                var temp = list[i+1];
-                list[i+1] = list[i];
-                list[i] = temp;
-                if (i > 0)
-                    i = i - 2
-            }
-        }
-
-        return changed;
-    },
-
-    convertArrayToSupplierList: function(array, format) {
-        var list = [];
-        for (var i = 0; i < array.length; i++) {
-            list.push({ value: new FormatDef.constructor(array[i], format) });
-        }
-        return list;
-    },
-
-    clone: function(object) {
-        return JSON.parse(JSON.stringify(object));
-    },
-
-    findDifferences: function(from, to) {
-        var j = 0;
-
-        var obj = { removed: [], added: [] };
-
-        if ((from === null || _.isUndefined(from)) && (to === null || _.isUndefined(to)))
-            return obj;
-        else if (from === null || _.isUndefined(from)) {
-            for (j = 0; j < to.length; j++)
-                obj.added.push(to[j]);
-        }
-        else if (to === null || _.isUndefined(to)) {
-            for (j = 0; j < from.length; j++)
-                obj.removed.push(from[j]);
-        }
-        else {
-            for (j = 0; j < from.length; j++) {
-                if (this.listContains(to, from[j]) === false)
-                    obj.removed.push(from[j]);
-            }
-
-            for (j = 0; j < to.length; j++) {
-                if (this.listContains(from, to[j]) === false)
-                    obj.added.push(to[j]);
-            }
-        }
-
-        return obj;
-    },
-
-    listContains: function(list, value) {
-        for (var i = 0; i < list.length; i++) {
-            if (FormatDef.variable.isEqual(list[i], value))
-                return true;
-        }
-
-        return false;
-    },
-
-    containsCovariate: function(value, covariates) {
-        for (var i = 0; i < covariates.length; i++) {
-            if (FormatDef.variable.contains(value, covariates[i]))
-                return true;
-        }
-
-        return false;
-    }
 });
 
-module.exports = { ui : ancovaLayout, options: options, customControls: controls };
+module.exports = { ui : ancovaLayout, actions: actions, options: options };
