+++ jamovi/ui/loglinear.src.js
@@ -2,6 +2,7 @@
 'use strict';
 
 var options = require("./loglinear.options");
+var actions = require("./loglinear.actions");
 
 var layout = ui.extend({
 
@@ -22,6 +23,7 @@ var layout = ui.extend({
                     label: "Counts (optional)",
                     showColumnHeaders: false,
                     maxItemCount: 1,
+                    itemDropBehaviour: "overwrite",
                     columns: [
                         { type: "listitem.variablelabel", name: "column1", label: "", format: FormatDef.variable, stretchFactor: 1 }
                     ]
@@ -87,204 +89,6 @@ var layout = ui.extend({
         }
     ],
 
-    actions: [
-        {
-            onChange: "ci", execute: function(ui) {
-                ui.ciWidth.setEnabled(ui.ci.value());
-            }
-        },
-        {
-            onEvent: "view.data-initialising", execute: function(ui) {
-                this._lastVariableList = null;
-                this._lastCurrentList = null;
-                this._initialising = true;
-            }
-        },
-        {
-            onChange: "factors", execute: function(ui) {
-                this.calcModelTerms(ui);
-            }
-        },
-        {
-            onChange: "modelTerms", execute: function(ui) {
-                this.filterModelTerms(ui);
-            }
-        },
-        {
-            onEvent: "view.data-initialised", execute: function(ui) {
-                if (this._lastVariableList === null)
-                    this.calcModelTerms(ui);
-
-                if (this._lastCurrentList === null)
-                    this.filterModelTerms(ui);
-
-                this._initialising = false;
-            }
-        }
-    ],
-
-    clone: function(object) {
-        return JSON.parse(JSON.stringify(object));
-    },
-
-    findDifferences: function(from, to) {
-        var j = 0;
-
-        var obj = { removed: [], added: [] };
-
-        if ((from === null || _.isUndefined(from)) && (to === null || _.isUndefined(to)))
-            return obj;
-        else if (from === null || _.isUndefined(from)) {
-            for (j = 0; j < to.length; j++)
-                obj.added.push(to[j]);
-        }
-        else if (to === null || _.isUndefined(to)) {
-            for (j = 0; j < from.length; j++)
-                obj.removed.push(from[j]);
-        }
-        else {
-            for (j = 0; j < from.length; j++) {
-                if (this.listContains(to, from[j]) === false)
-                    obj.removed.push(from[j]);
-            }
-
-            for (j = 0; j < to.length; j++) {
-                if (this.listContains(from, to[j]) === false)
-                    obj.added.push(to[j]);
-            }
-        }
-
-        return obj;
-    },
-
-    listContains: function(list, value) {
-        for (var i = 0; i < list.length; i++) {
-            if (FormatDef.variable.isEqual(list[i], value))
-                return true;
-        }
-
-        return false;
-    },
-
-    sortByLength : function(list) {
-        var changed = false;
-        for (var i = 0; i < list.length; i++) {
-            var l1 = 1;
-            if (Array.isArray(list[i]))
-                l1 = list[i].length;
-
-            var l2 = 1;
-            if (Array.isArray(list[i+1]))
-                l2 = list[i+1].length;
-
-            if (list.length > i + 1 && (l1 > l2)) {
-                changed = true;
-                var temp = list[i+1];
-                list[i+1] = list[i];
-                list[i] = temp;
-                if (i > 0)
-                    i = i - 2
-            }
-        }
-
-        return changed;
-    },
-
-    calcModelTerms: function(ui) {
-        var variableList = this.clone(ui.factors.value());
-        if (variableList === null)
-            variableList = [];
-
-        ui.modelSupplier.setValue(this.convertArrayToSupplierList(variableList, FormatDef.variable));
-
-        var diff = { removed: [], added: [] };
-        if (this._lastVariableList !== null)
-            diff = this.findDifferences(this._lastVariableList, variableList);
-        this._lastVariableList = variableList;
-
-        if (this._initialising)
-            return;
-
-        var currentList = this.clone(ui.modelTerms.value());
-        if (currentList === null)
-            currentList = [];
-
-        for (var i = 0; i < diff.removed.length; i++) {
-            for (var j = 0; j < currentList.length; j++) {
-                if (FormatDef.variable.contains(currentList[j], diff.removed[i])) {
-                    currentList.splice(j, 1);
-                    j -= 1;
-                }
-            }
-        }
-
-        if (currentList === null)
-            currentList = [];
-
-        for (var i = 0; i < diff.added.length; i++) {
-            var listLength = currentList.length;
-            for (var j = 0; j < listLength; j++) {
-                var newVar = currentList[j];
-                if (Array.isArray(newVar))
-                    newVar = this.clone(newVar);
-                else
-                    newVar = [newVar];
-                newVar.push(diff.added[i])
-                currentList.push(newVar);
-            }
-            currentList.push(diff.added[i]);
-        }
-
-        ui.modelTerms.setValue(currentList);
-    },
-
-    filterModelTerms : function(ui) {
-        var currentList = this.clone(ui.modelTerms.value());
-        if (currentList === null)
-            currentList = [];
-
-        var diff = { removed: [], added: [] };
-        if (this._lastCurrentList !== null)
-            diff = this.findDifferences(this._lastCurrentList, currentList);
-        this._lastCurrentList = currentList;
-
-        if (this._initialising)
-            return;
-
-        var changed = false;
-        if (diff.removed.length > 0 && currentList !== null) {
-            var itemsRemoved = false;
-            for (var i = 0; i < diff.removed.length; i++) {
-                var item = diff.removed[i];
-                for (var j = 0; j < currentList.length; j++) {
-                    if (FormatDef.variable.contains(currentList[j], item)) {
-                        currentList.splice(j, 1);
-                        j -= 1;
-                        itemsRemoved = true;
-                    }
-                }
-            }
-
-            if (itemsRemoved)
-                changed = true;
-        }
-
-        if (this.sortByLength(currentList))
-            changed = true;
-
-        if (changed)
-            ui.modelTerms.setValue(currentList);
-    },
-
-    convertArrayToSupplierList: function(array, format) {
-        var list = [];
-        for (var i = 0; i < array.length; i++) {
-            list.push({ value: new FormatDef.constructor(array[i], format) });
-        }
-        return list;
-    },
-
-
 });
 
-module.exports = { ui : layout, options: options };
+module.exports = { ui : layout, actions: actions, options: options };
